OP-1216 Adapt GUI to DECIMAL/INT money and quantity types (#2212)
* OP-1216 Adapt GUI to DECIMAL/INT money and quantity types

Adapt the Swing GUI to the core type change (OP-1216):
- Accounting totals (Balance/Payments/User) and PatientBillEdit use the
  BigDecimal amount/balance getters directly.
- Price model/browser handle BigDecimal price (display narrows to double
  only for rendering); MedicalEdit and WardPharmacy screens build
  movements with BigDecimal/int quantities.
- Tests updated for the new types.

* OP-1216 Allow fractional quantities when dispensing ward drugs

The ward stock movement quantity (MMVN_MDSR_QTY) is DECIMAL per OP-1216
("could be half unit"), but WardPharmacyNew.askQuantity parsed the input
with Integer.parseInt, so a half-unit dispense (e.g. 2.5) was rejected
with "Insert a valid value". Switch the charging quantity flow to double
(askQuantity / automaticChoose / chooseLot / addItem and the quantity
column type) so fractional ward movements can actually be entered; the
movement is still stored as BigDecimal. The incoming ward stock quantity
(MDSRWRD_IN_QTI) stays integer, as defined by the ticket.

* OP-1216 Fix NPE rendering the price list tree with group rows

PriceModel.getValueAt called price.getPrice().doubleValue() for the price
column, but group/category nodes in the price tree have a null price, so
the whole Price Lists screen failed to render (empty tree, repeated
NullPointerException). Guard against a null price (group rows show an
empty price cell, as before), keeping doubleValue() for actual prices.

* OP-1216 Import BigDecimal instead of using the fully-qualified name in tests

* OP-1216 Add validation message for whole-unit ward transfers

Message shown when a ward-to-ward transfer is entered with a fractional
quantity, which core now rejects (the incoming side is stored as an integer).

* Fix exception dialog to get actual error message

* Fix CI core branch checkout for PR forks

* OP-1216 Compute ward pharmacy quantities in BigDecimal

Following review (mwithi): parse and do the ward-movement quantity arithmetic in
WardPharmacyNew with BigDecimal instead of double/int — the dispense quantity is
parsed with new BigDecimal(input) and the running totals, comparisons and stock
checks use compareTo/signum/subtract, so no calculation goes through floating point.
WardPharmacy and WardPharmacyRectify are adapted to the BigDecimal MedicalWard.getQty().
Verified end-to-end on a demo DB: a 2.5 dispense is stored exactly and a 2.5 ward-to-ward
transfer is rejected as non-integer.

* OP-1216 Handle BigDecimal inventory quantities in the GUI

- InventoryWardEdit: the ward inventory now accepts fractional real
  quantities (quantity columns switched to BigDecimal with the existing
  decimal renderer)
- InventoryEdit: adapted to the BigDecimal entity while keeping integer
  entry, consistent with the integer main-store stock
- WardPharmacyRectify: rectification arithmetic in BigDecimal, spinner
  values normalized to Double, cost input parsed without a double round-trip

* OP-1216 Reject fractional real quantities in the main-store inventory

The inventory table uses a free-text cell editor, so a fractional quantity
reached the model and was persisted while the integer display truncated it.
Main-store stock is integer: reject fractional input with the standard
invalid-quantity error, like negative values.

* OP-1216 Tidy inventory row creation and main-store integer display

- Extract the main-store quantity into a local instead of converting it
  twice per constructor call, keeping the lines within 160 columns
- Render the theoretical quantity as an integer in the main-store
  inventory, matching the real-quantity column

* OP-1216 Add the i18n message for fractional inventory adjustments

* OP-1216 Rename the fractional-quantities message to match the wider validation

---------

+++ src/test/java/org/isf/accounting/gui/TestPayment.java
@@ -21,6 +21,8 @@
  */
 package org.isf.accounting.gui;
 
+import java.math.BigDecimal;
+
 import org.isf.accounting.model.Bill;
 import org.isf.accounting.model.BillPayments;
 
@@ -28,7 +30,7 @@ public class TestPayment {
 
 	public static BillPayments withAmountAndBill(double amount, Bill bill) {
 		BillPayments billPayments = new BillPayments();
-		billPayments.setAmount(amount);
+		billPayments.setAmount(BigDecimal.valueOf(amount));
 		billPayments.setBill(bill);
 		return billPayments;
 	}
